requirements/common.txt
fastapi==0.88.0
uvicorn==0.20.0
python-multipart==0.0.5
SQLAlchemy==1.4.45
asyncpg==0.27.0
passlib==1.7.4
python-jose==3.3.0
pydantic[dotenv]==1.10.2
bcrypt==4.0.1
cryptography==38.0.4
email-validator==1.3.0
fastapi_pagination==0.11.1
redis==4.4.0
psycopg2==2.9.5
sentry-sdk==1.11.1